Move ginger template to Setup.hs
