+++ wlhs-bindings.cabal
@@ -8,8 +8,18 @@ author:          Brad Neimann
 tested-with:     ghc ==9.4.8
 synopsis:        Bindings to wlroots (and libwayland)
 category:        Graphics
-build-type:      Simple
+build-type:      Custom
 extra-doc-files: CHANGELOG.md
+
+custom-setup
+    setup-depends:
+        base >=4.16.4.0 && <5,
+        Cabal >=3.10.1.0 && <3.11,
+        directory >=1.3 && <1.4,
+        ginger >=0.10 && <0.11,
+        mtl >=2.2 && <2.3,
+        text >=1.2 && <1.3,
+        unordered-containers
 
 library
     exposed-modules:
@@ -34,5 +44,7 @@ library
     hs-source-dirs:     src
     default-language:   Haskell2010
     default-extensions: CApiFFI
-    ghc-options:        -Wall -fno-show-valid-hole-fits -optc -Wno-discarded-qualifiers
+    ghc-options:
+        -Wall -fno-show-valid-hole-fits -optc -Wno-discarded-qualifiers
+
     build-depends:      base >=4.16.4.0 && <5
